initial commit
this python file, as it is in this initial commit, is a copy of the
pnexpose.py taken from the Rapid7 Splunk App.  We will likely need to
modify this for internal use.  By commiting the file as it was in the
Splunk app, before changes (other than file name and changing tabs >
spaces) we'll have a record of those modifications.